Added two command classes that need finishing
Added AcceptDuelCommand and DuelCommand classes. They are not functional yet. I am trying to get a skeleton going so I can create the fundamentals of the plugin.

Added the Invites class. I need to figure out how to handle data like that where it needs to be readily accessible but only in the ram, not a file.