Added handling for the rock command
Need to figure out how to handle the rock/paper/scissor commands and whether or not it should be done from one class or have a class for each.

The class needs to be able to communicate with the game based on what the command class returns. The default seems to be a boolean.