Removed unnecessary code and made round data persist
The plugin now stores the GameInstance on the main thread and updates it whenever a player in that game instance chooses rock, paper or scissors. It displays the scores now and says you win if you reach a score of 3. It is still not a functional game, but it's progress!